+++ Src/SimSession.h
@@ -8,12 +8,12 @@
 class SimSession
 {
 public:
-  SimSession(uint64_t id, const String& name);
+  SimSession(uint32_t id, const String& name);
 
   bool_t start(const String& engine, double balanceBase, double balanceComm);
 
 private:
-  uint64_t id;
+  uint32_t id;
   String name;
   Process process;
 };
